% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_total.R
\name{add_total}
\alias{add_total}
\title{Append a totals row and/or column to a data.frame/ tibble.}
\usage{
add_total(dat, where = "row", fill = "-", na.rm = TRUE,
  name = "Total")
}
\arguments{
\item{name}{}
}
\description{
This function use the first column of the input data.frame. Non-numeric columns are converted to character class and have a user-specified fill character inserted in the totals row.
}
