% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anosim_plot.R
\name{anosim_plot}
\alias{anosim_plot}
\title{ANOSIM plot}
\usage{
anosim_plot(anosim_object)
}
\arguments{
\item{comm}{anosim object}
}
\value{
ggplot2
}
\description{
Analysis of similarities (ANOSIM) provides a way to test statistically whether there is a significant difference between two or more groups of sampling units.
}
\examples{
library(vegan)
data(dune)
data(dune.env)
dune.dist <- vegdist(dune)
dune.ano <- with(dune.env, anosim(dune.dist, Management))
summary(dune.ano)
plot(dune.ano)
anosim_plot(dune.ano)
}
