% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_distbl.R
\name{as_distbl}
\alias{as_distbl}
\title{as_distbl}
\usage{
as_distbl(x, diagonal = NA)
}
\arguments{
\item{x}{dist object}

\item{diagonal}{Set as NA character by default}
}
\value{
tibble
}
\description{
A function to coerce object of class dist
(mainly created as the output of different packages used
for ecological diversity measurings. for instance:
stats:: dist(), vegan:: vegdist(), betapart::beta.pair()...)
to a tibble object.
}
\keyword{internal}
