% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_distdf.R
\name{as_distdf}
\alias{as_distdf}
\title{as_tbldist}
\usage{
as_distdf(x)
as_distdf(x, type = "long")
}
\arguments{
\item{x}{Object class dist.}

\item{type}{select the output structure: "wide", for square data_frame; "long", for tidy data_frame.}
}
\value{
a tibble object.
}
\description{
A function to coerce objects of class dist
(mainly created as the output of different packages used
for ecological diversity measurings. for instance:
stats:: dist(), vegan:: vegdist(), betapart::beta.pair()...)
to a data frame.
}
\examples{
require(vegan)
data("dune")
vegdist(dune) \%>\%
as_distdf()
require(betapart)
beta <- beta.pair.abund(dune)
beta$beta.bray.bal \%>\%
as_distdf(type = "long")

}
