% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_tbldist.R
\name{as_tbldist}
\alias{as_tbldist}
\title{as_tbldist}
\usage{
as_distdf(x, diagonal = NA)
}
\arguments{
\item{x}{Object class dist.}
}
\value{
a tibble object.
}
\description{
A function to coerce objects of class dist
(mainly created as the output of different packages used
for ecological diversity measurings. for instance:
stats:: dist(), vegan:: vegdist(), betapart::beta.pair()...)
to a data frame.
}
\examples{
require(vegan)
data("dune")
vegdist(dune) \%>\%
as_tbldist()

}
