% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bci.R
\docType{data}
\name{bci}
\alias{bci}
\title{BCI data modified}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 503 rows and 4 columns.
}
\usage{
bci
}
\description{
BCI data modified
}
\section{Slots}{

\describe{
\item{\code{data}}{A data frame}
}}

\keyword{datasets}
