% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comdf_to_tbl.R
\name{comdf_to_tbl}
\alias{comdf_to_tbl}
\title{comdf_to_tbl}
\usage{
comdf_to_tbl(comm, type = "num")
}
\arguments{
\item{comm}{Community data, a matrix-like object.}

\item{type}{Class of the name string for each site; "num" - numeric id or "char" - character string id.}
}
\description{
This function converts a community data.frame to a tidy tibble object.
}
\examples{
require(vegan)
data("dune")
dune \%>\%
comdf_to_tbl()
}
