% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dune.R
\docType{data}
\name{dune}
\alias{dune}
\title{Vegetation and Environment in Dutch Dune Meadows}
\format{
code dune is a data frame of observations of 30 species at 20
sites. The species names are abbreviated to 4+4 letters (see
codelinkmake.cepnames). The following names are changed from
the original source (Jongman et al. 1987): emphLeontodon
autumnalis to emphScorzoneroides, and emphPotentilla
palustris to emphComarum.
}
\source{
Jongman, R.H.G, ter Braak, C.J.F & van Tongeren,
O.F.R. (1987). emphData Analysis in Community and Landscape
Ecology. Pudoc, Wageningen.
}
\usage{
data(dune)
}
\description{
The dune meadow vegetation data, codedune, has cover class values
of 30 species on 20 sites. The corresponding environmental data frame
codedune.env has following entries:
}
\examples{
  data(dune)

}
\keyword{datasets}
