% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/long_distbl.R
\name{long_distbl}
\alias{long_distbl}
\title{long_distbl}
\usage{
long_distbl(df, x, index)
}
\arguments{
\item{df}{distbl object}

\item{x}{name(for instance: plot, site, location...)}

\item{index}{name of the distance index used}
}
\value{
tibble
}
\description{
Convert dist tibble object to long format tibble.
}
\examples{
require(vegan)

data("dune")

vegdist(dune) \%>\%
as_distbl() \%>\%
long_distbl(plot, dist_index)

}
