% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rankabund.R
\name{rankabund}
\alias{rankabund}
\title{Rankabun: Rank abundance}
\usage{
rankabund(comm, group = "none")
}
\arguments{
\item{comm}{Community data.}

\item{group}{Index for changing the output:
"none" the default output.
"sites" the rank abundance indexes are estimated for each row (plot, site).}
}
\value{
a tibble
}
\description{
Calculating rank-abundance curves data.
}
\examples{

require(vegan)
data(dune)
rankabund(dune)
rankabund(dune, group = "sites")

require(fossil)
data("fdata.mat")
rankabund(t(fdata.mat))
rankabund(t(fdata.mat), group = "sites")
}
