% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rankabun_df.R
\name{rankabund_df}
\alias{rankabund_df}
\title{Rankabun: Rank abundance}
\usage{
rankabund_df(comm, group = "none")
}
\arguments{
\item{comm}{Community data.}

\item{group}{Index for changing the output:
"none" the default output.
"sites" the rank abundance indexes are
        estimated for each row (plot, site).}
}
\value{

}
\description{
Calculating rank-abundance curves data.
}
\examples{

require(vegan)
data(dune)
rankabun_df(dune)
rankabun_df(dune, group = "sites")

require(fossil)
data("fdata.mat")
rankabun_df(t(fdata.mat))
rankabun_df(t(fdata.mat), group = "sites")
}
