% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rankabun_plot.R
\name{rankabund_plot}
\alias{rankabund_plot}
\title{rankabun_plot}
\usage{
rankabund_plot(comm, method = "abundance")
}
\arguments{
\item{comm}{A community data.frame.}

\item{method}{Options are "abundance" and "logabun".}
}
\value{
A ggplot object.
}
\description{
rankabun_plot
}
\examples{
require(vegan)
data(dune)
rankabund_plot(dune)
rankabund_plot(dune, method ="logabun")
}
