% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rarecurve_plot.R
\name{rarecurve_plot}
\alias{rarecurve_plot}
\title{rarecurve_plot}
\usage{
rarecurve_plot(comm, color = "int")
}
\arguments{
\item{comm}{Community data a matrix}

\item{color}{Options  "int" and "id".}
}
\value{

}
\description{
Function rarecurve_plot draws a rarefaction curve for each row of the input data.
}
\details{
color "int" display the intinsity based on species abundance, and "id" display unique colour for each site.
}
\examples{
require(vegan)
data("dune")
rarecurve_plot(dune)
rarecurve_plot(dune, facet.var = "sites")

}
