% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specaccumplot.R
\name{specaccumplot}
\alias{specaccumplot}
\title{specaccumplot}
\usage{
specaccumplot(comm, method = "exact")
}
\arguments{
\item{comm}{Community data frame.}

\item{method}{Species accumulation method (partial match).
 Use the same options as specaccum function from vegan package,
"random" adds sites in random order, "exact" finds the
 expected (mean) species richness, "coleman" finds the expected
 richness, and "rarefaction" finds the mean when accumulating
 individuals instead of sites. Except for "collector" method.}
}
\value{

}
\description{
specaccumplot
}
\examples{

require(vegan)
data(dune)
specaccumplot(dune)

}
\seealso{
[specaccum()] from vegan.
}
