% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_to_comm.R
\name{tbl_to_comm}
\alias{tbl_to_comm}
\title{Tibble to community data.frame}
\usage{
tbl_to_comm(df, site, taxon, abun)
}
\arguments{
\item{df}{tibble for community richnes data}

\item{site}{variable for group species (site, plot).}

\item{taxon}{species column name.}

\item{abun}{richnes species column.}
}
\value{
a community matrix
}
\description{
Tibble to community data.frame
}
