% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_comm.R
\name{to_comm}
\alias{to_comm}
\title{Tibble to community data.frame}
\usage{
to_comm(df, x, y, z)
}
\arguments{
\item{df}{tibble for community richnes data}

\item{x}{variable for group species (site, plot).}

\item{y}{species column name.}

\item{z}{richnes species column.}
}
\value{

}
\description{
Tibble to community data.frame
}
\examples{
df \%>\%
to_comm(x, y, z)

}
