% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anosim_plot.R
\name{anosim_plot}
\alias{anosim_plot}
\title{ANOSIM plot}
\usage{
anosim_plot(comm)
}
\arguments{
\item{comm}{anosim object}
}
\value{
ggplot2
}
\description{
Analysis of similarities (ANOSIM) provides a way to test statistically whether there is a significant difference between two or more groups of sampling units.
}
\keyword{internal}
