% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comm_to_tbl.R
\name{comm_to_tbl}
\alias{comm_to_tbl}
\title{Community data.frame to tibble.}
\usage{
comm_to_tbl(comm)
}
\arguments{
\item{comm}{Community data.}
}
\value{

}
\description{
This function converts a community data.frame to a tidy tibble object.
}
\examples{
require(vegan)
data("dune")
dune \%>\%
comm_to_tbl()
}
