% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ivi_index.R
\name{ivi_index}
\alias{ivi_index}
\title{ivi : Importance Valor Index}
\usage{
ivi_index(df, sample_unit, taxon, basal_area)
}
\arguments{
\item{df}{A data_frame, with the information for each tree in the census.}

\item{sample_unit}{Name of the variable for the minimum unit sample.}

\item{taxon}{Name of the variable (column) that contein species name.}

\item{basal_area}{Name of the variable (column) that contein basal area.}
}
\value{
a tibble
}
\description{
ivi : Importance Valor Index
}
\examples{
require(diversity.tool)
data("bci")
bci |>
ivi_index(sub_plot, species, area_basal)

}
