% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ivi_index.R
\name{ivi_index}
\alias{ivi_index}
\title{ivi : Importance Valor Index}
\usage{
ivi_index(df, sp = "species", ab = "area_basal", group = "sub_plot")
}
\arguments{
\item{df}{A data_frame, with the information for each tree in the census.}

\item{sp}{Name of the variable (column)that contein species name.}

\item{ab}{Name of the variable (column) that contein basal area.}

\item{group}{Name of the variable for the minimum unit sample.}
}
\value{

}
\description{
ivi : Importance Valor Index
}
\examples{
require(diversity.tool)
data("bci")
bci \%>\%
ivi(sp = "species", ab = "area_basal", group = "sub_plot")
}
