% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rarecurve_plot.R
\name{rarecurve_plot}
\alias{rarecurve_plot}
\title{rarecurve_plot}
\usage{
rarecurve_plot(comm, facet.var = "none")
}
\arguments{
\item{comm}{Community data a matrix}

\item{facet.var}{Option for plot output, default is "none" and "sites".}
}
\description{
Function rarecurve_plot draws a rarefaction curve for each row of the input data.
}
\examples{
require(vegan)
data("dune")
rarecurve_plot(dune)
rarecurve_plot(dune, facet.var = "sites")

}
