% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_richg.R
\name{sp_richg}
\alias{sp_richg}
\title{Richness and frequency by treatment.}
\usage{
sp_richg(comm1, char, factor = "")
}
\arguments{
\item{comm1}{community data matrix.}

\item{char}{characteristic from the community.}

\item{factor}{name of the variable used for build a grouped community data.}
}
\description{
The function returns a tibble: "n_plot" the frequency of grouped variable. "n_species" species richness of the grouped variable.
}
\examples{
require(dune)
data("dune")
data("dune.env")
sp_richg(dune, dune.env, factor = "Management")
}
